Add a command line interface with structopt
This crate is split up into a library and two binaries:

- fxrunner: the binary responsible for running Firefox on the reference
  hardware;
- fxrecorder: the binary responsible for doing frame capture; and
- fxrecord: the library with common code for both fxrunner and
  fxrecorder.

The configuration for both fxrunner and fxrecorder is parsed from a TOML
file (configurable via the `--config` command-line option). For ease of
development, the fxrunner and fxrecorder configuration are both able to
be included in a single config file (`fxrecord.toml`) using sections. An
example configurationf ile (`fxrecord.example.toml`) is provided.