Use the Windows shutdown implementation when restarting
A `Shutdown` trait has been added which allows for restarting the
system. The main implementation of the `Shutdown` trait is the
`WindowsShutdown` implementation, which delegates directly to the
Windows API. A debug-build only version, `DebugShutdown`, only actually
shuts down the system if the `--skip-restart` flag is not provided
(which is also only accepted on debug builds). This exists as a trait
mainly so that the various forms of handshaking can be tested.