Use TryFrom impl for Handle instead of From
`Handle` is now only constructible through a `TryFrom` impl, which
checks against the invalid handle value and returns the last Windows
error if it matches.

This also corrects the error checking we were doing while acquiring a
handle to the logical C:\ drive, as the invalid handle value is not a
null pointer.