Use TermDecorator for logging
Using `PlainDecorator` and passing it a copy of `std::io::stdout()`
results in it locking the stdout mutex every time we log, which is what
we were trying to avoid by using slog_async. Now we use the
`TermDecorator`, which uses the `term` crate to talk to the underlying
file handle directly, avoiding locking.